requirements.txt
canvasapi==2.2.0
python_dateutil==2.8.2
pytz==2022.1
requests==2.28.0
